user_auth/requirements.txt
alembic==1.16.1
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
greenlet==3.2.1
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
